llama_model_size is explicity sized as uint64_t
this is an implicit cast

+++ include/llama.h
@@ -469,7 +469,7 @@ extern "C" {
     LLAMA_API int32_t llama_model_desc(const struct llama_model * model, char * buf, size_t buf_size);
 
     // Returns the total size of all the tensors in the model in bytes
-    LLAMA_API size_t llama_model_size(const struct llama_model * model);
+    LLAMA_API uint64_t llama_model_size(const struct llama_model * model);
 
     // Returns the total number of parameters in the model
     LLAMA_API uint64_t llama_model_n_params(const struct llama_model * model);
